✨  Add: Firebase Appを初期化する実装
