+++ src/@types/piledit.ts
@@ -340,27 +340,3 @@ export class PTab implements PTabIF {
 export type PTabs = {
   [key: string]: PTab;
 }
-
-export interface ConfigIF {
-  apiKey?: string;
-  authDomain?: string;
-  storageBucket?: string;
-  databaseURL?: string;
-  projectId?: string;
-  messagingSenderId?: string;
-}
-
-export interface CredentialsIF {
-  config: ConfigIF;
-}
-
-export const credentials: CredentialsIF = {
-  config: {
-    apiKey: process.env.VUE_APP_FIREBASE_API_KEY,
-    authDomain: process.env.VUE_APP_FIREBASE_AUTH_DOMAIN,
-    storageBucket: process.env.VUE_APP_FIREBASE_STORAGE_BUCKET,
-    databaseURL: process.env.VUE_APP_FIREBASE_DB_URL,
-    projectId: process.env.VUE_APP_FIREBASE_PROJECT_ID,
-    messagingSenderId: process.env.VUE_APP_FIREBASE_MESSAGING_SENDER_ID
-  }
-}
