+++ src/firebase/auth.ts
@@ -1,0 +1,5 @@
+import { App } from './app'
+import firebase from 'firebase'
+import 'firebase/auth'
+
+export const Auth: firebase.auth.Auth = App.auth()
