+++ src/firebase/credentials.ts
@@ -1,0 +1,12 @@
+import { CredentialsIF } from './types/credentials'
+
+export const credentials: CredentialsIF = {
+  config: {
+    apiKey: process.env.VUE_APP_FIREBASE_API_KEY,
+    authDomain: process.env.VUE_APP_FIREBASE_AUTH_DOMAIN,
+    storageBucket: process.env.VUE_APP_FIREBASE_STORAGE_BUCKET,
+    databaseURL: process.env.VUE_APP_FIREBASE_DB_URL,
+    projectId: process.env.VUE_APP_FIREBASE_PROJECT_ID,
+    messagingSenderId: process.env.VUE_APP_FIREBASE_MESSAGING_SENDER_ID
+  }
+}
